feature: get data reorganized in a Service & Hook
