+++ pruebas/01-reading-list/foleas/src/services/getData.ts
@@ -1,0 +1,7 @@
+import { Book, Library } from "../types";
+
+export const getData = async (endpoint: string): Promise<Array<Book>> => {
+  const response = await fetch(endpoint);
+  const data: Library = (await response.json()) as Library;
+  return data.library;
+};
